Correcciones al modelo de base de datos

+++ database/migrations/2026_06_23_202701_add_photo_to_users_table.php
@@ -11,11 +11,11 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-
-            $table->string('photo')->nullable()->after('email');
-
-        });
+        if (!Schema::hasColumn('users', 'photo')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->string('photo')->nullable()->after('email');
+            });
+        }
     }
 
     /**
@@ -23,10 +23,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-
-            $table->dropColumn('photo');
-
-        });
+        if (Schema::hasColumn('users', 'photo')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('photo');
+            });
+        }
     }
 };
