+++ database/migrations/2026_08_23_184046_add_firma_to_users_table.php
@@ -8,16 +8,20 @@ return new class extends Migration
 {
     public function up(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->string('firma')
-                ->nullable();
-        });
+        if (!Schema::hasColumn('users', 'firma')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->string('firma')
+                    ->nullable();
+            });
+        }
     }
 
     public function down(): void
     {
-        Schema::table('users', function (Blueprint $table) {
-            $table->dropColumn('firma');
-        });
+        if (Schema::hasColumn('users', 'firma')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('firma');
+            });
+        }
     }
 };
